+++ barzer_shell.cpp
@@ -231,7 +231,7 @@ static int bshf_spell( BarzerShell* shell, char_cp cmd, std::istream& in )
 
 			std::cerr << "misspelling correctin it. got " << scResult.second << " suggestions\n";
 			const char*const* sugg = spellChecker.getAllSuggestions();
-			for( int i =0; i< scResult.second; ++i ) {
+			for( size_t i =0; i< scResult.second; ++i ) {
 				std::cerr << sugg[i] << std::endl;
 			}
 		} else {
